\name{startEPSDevice}
\alias{startEPSDevice}
\title{Prepare for printing}
\description{
    Calculate the expected number of transect sections occupied by a species for a bbs transect
}
\usage{
  startEPSDevice(name,w=10,h=8)
}
\arguments{
  \item{name}{Filename for printing (without eps extension)}
  \item{w}{Width of the eps bounding box}
  \item{h}{Height of the eps bounding box} 
}
\value{
    Used for it's side effects
}
\seealso{
  \code{\link{printBarplot}}.
}
#\examples{}
\keyword{}
